add HeroImage component
I've been copying the markup between posts for a while and I want
to make this a generic component. The hdrImage is a mistake, it's an
absolute catastrophe and I want to see if it causes more engagement
with the posts.

Automated projection from SDR to HDR is a very hit or miss affair.
Most of the time, it will miss. Sometimes however, it will work fairly
well. Well enough that it's worth experimenting with people's retinas
in order to see how bad of an idea this actually is.

Science is not about "why". It's about "why not?"

Signed-off-by: Xe Iaso <xe@tigrisdata.com>